backend/requirements.txt
bcrypt==5.0.0
blinker==1.8.2
click==8.1.8
colorama==0.4.6
flask==3.0.3
Flask-Cors==5.0.0
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
werkzeug==3.0.6
zipp==3.20.2
reportlab<4.0.0
typing-extensions==4.13.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
cloudinary
